Test model in API
