+++ Med_Cabinet/requirements.txt
@@ -1,0 +1,16 @@
+pandas
+numpy
+requests
+flask
+flask_restful
+json
+pickle-mixin
+sklearn.feature_extraction.text
+re
+string
+os
+sys
+dotenv
+psycopg2
+
+spacy_install_virtualenv
